refactor!: EvaluationContextConfigurer classes
The EvaluationContextConfigurer classes, DefaultEvaluationContextConfigurer and NoOpEvaluationContextConfigurer, previously located in the org.wickedsource.docxstamper.el package, have been moved into the pro.verron.officestamper.preset package. The primary motive of this refactor was to streamline the code and reduce package dependence. DocxStamperConfiguration has been updated to reflect this change.
